+++ packages/vue/src/interface.ts
@@ -1,8 +1,15 @@
 import Vue from 'vue';
-import { CuriConfig } from '@curi/core';
+import { CuriConfig, Response } from '@curi/core';
+import { Action } from '@hickory/root';
+
+export interface CuriPluginObject {
+  config: CuriConfig;
+  response: Response;
+  action: Action;
+}
 
 declare module 'vue/types/vue' {
   interface Vue {
-    $curi: CuriConfig;
+    $curi: CuriPluginObject;
   }
 }