+++ packages/vue/types/interface.d.ts
@@ -1,6 +1,12 @@
-import { CuriConfig } from '@curi/core';
+import { CuriConfig, Response } from '@curi/core';
+import { Action } from '@hickory/root';
+export interface CuriPluginObject {
+    config: CuriConfig;
+    response: Response;
+    action: Action;
+}
 declare module 'vue/types/vue' {
     interface Vue {
-        $curi: CuriConfig;
+        $curi: CuriPluginObject;
     }
 }