+++ packages/vue/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Next
+
+* Add `active` support to the `<curi-link>`, where `active` is an object with a `merge` function and an optional `partial` boolean.
+* Expand the `$curi` object to contain `response` and `action` (`this.$curi = { config, response, action }`).
+
 ## 1.0.0-beta.8
 
 * `<curi-link>` uses `history.navigate`, not `history.update`.